requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.29
pymysql==1.1.0
python-dotenv==1.0.1